Added minimisation, Event Display

+++ src/hit.h
@@ -15,7 +15,7 @@ struct hit{
     
     uint16_t layer; // lowest order 3 bits of a 16 bit datum
     uint16_t wire; // wire number within the layer, next 3 bits
-    uint16_t TDC; // final 10 bits
+    double TDC; // final 10 bits
     
 };
